Changes to toggle button subscribers signature v0.3.3
Now, toggle button subscribers and normal button subscribers have the same
signature:

  function (event, eventData) {}

The key difference is that if it is infact a toggle button, the eventData
object will have a 'state' property that will reflect the state of the button
when it was fired.  Eg. if the user clicked on a less link, the value of state
will be 'less'.

Also, condensed event handling to a single point rather than the two point, and
rather confusing approach adopted earlier.

Thanks to @stoewer for pointing all of it out.